config: add collectors argument to collect command
"collect" command needs a way to specify what collectors to use.

There's a slight complication on how to implement this. On the one
hand, we want to dynamically register collectors and clap's
ValueParser API does not allow lazy initialization. This would be solved
by creating the argument after all the registration phase is done.
However, if we do that, this argument will appear at the bottom of the
help string in the help heading of the last collector.

So, to work around this situation we declare the argument at the
begining and modify it after registration.

For this, and other, reasons we cannot just use clap's declare API but
we still want it to be part of CollectArgs, so manually populate the
field on match parsing.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>