probe: kernel: Add fexit support.
fexit is similar to raw_tracepoints in that ksym is set externally by
setting rodata section before loading the program.

The return code is appended to arguments.

A flag is added to ProbeConfig to indicate whether the probe supports
return value. That way we can ensure only the right hooks access it.

Signed-off-by: Adrián Moreno <amorenoz@redhat.com>