module: ovs: support filtering in userspace events
Currently filters are not affecting OVS USDT probes.

In kprobe:upcall_enqueue:
Do not use skb filter. We already filter in
raw_tracepoint:openvswitch:ovs_dp_upcall and only populate
inflight_upcall if needed.

Also, add the queue id to a new map called inflight_enqueue.

In usdt:recv_upcall: Lookup the queue id in inflight_enqueue map to know
if it was filtered or not. We still need do batch processing in order to
keep the number of recv_upcall and operations consistent. Therefore,
only skip sending the event if the upcall was filtered.

In op_flow_execute and op_flow_put: Like in recv_upcall, we need to do
batch processing unconditionally but we can skip the event if the upcall
is filtered.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>