module: ovs: use ovs_execute_actions to track..
action executions in ovs.

We were previously using execute_action_command which is run when
userspace has injected a packet.

This patch makes filtering difficult so using ovs_execute_actions
instead.

The kernel code goes like:
 - ovs_execute_actions(struct sk_buff *) (kprobe)
    - ovs_do_execute_action(struct sk_buff *, ...) (tracepoint)
    - ovs_do_execute_action(struct sk_buff *, ...) (tracepoint)
    - ...
 - ovs_execute_actions (kretprobe)

Filtering should happen on ovs_execute_actions mainly for two reasons:
 - It only happens once per call, i.e: more performant
 - It is called with the original skb. Some actions might change it.

This patch makes ovs_execute_actions kprobe insert the pid/tid into a
map to indicate that an action execution is in place (and the filter
passed). Currently filtering is not implemented, that will come in the
next patch.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>