Move the skb tracking logic to the core
We now handle all the tracking logic in the core, which will help for
filtering as well to take this into account. We still keep the tracking
event generation in its own module so it can be optional and so we can
see those events on probes where no other event is generated (when asked
by the user).

Signed-off-by: Antoine Tenart <atenart@redhat.com>