core: tracking: skb: handle partial frees
An skb data can be merged into another skb frags, and we'll never see
its data being freed. This is an issue as 1) this logic is often used in
TCP for optimization and 2) we won't be able to remove the corresponding
tracking information. This patch adds support for this.

A few things to note:

- We can use the same tracking id for the skb being merged and the one
  it is merged into as both lived before. And it's not easy to make the
  link only using the BPF side as multiple skbs might be merged into
  one.

- The only way we can track those events is by tracking
  kfree_skb_partial. But functions down the line are problematic to
  track as new tracking ids would be generated and not freed, which is
  two different issues. Luckily the only skb-trackable function is
  skb_release_head_state which is very low level and not really useful
  to track. Fixed this by adding a "no_tracking" flag to avoid adding
  tracking info for problematic probes.

Signed-off-by: Antoine Tenart <atenart@redhat.com>