events: split ebpf and json factories
Serde and ebpf event creation are combined into the EventSectionFactory.

However, the logic to actually parse the json object is in the Event
definition itself, not in the module which now only contains probe
loading logic.

For collection, it makes sense to dynamically enable collector modules,
but for unmarshaling json we cannot do that, we need to parse every
supported section.

Besides, the EventFactory is a trait implemented by both
FileEventFactory and BpfEventFactory but it doesn't seem to be a natural
fit for json parsing (i.e: FileEventFactory), e.g: we don't need to
start/stop the event creation.

All in all, it seems this attempt to unify an API for event unmarshaling
is not really fitting its purpose and is more of an opstacle.

This patch splits both event creation paths, simplifying event
unmarshaling (which now does not depend on modules).

Also, EventFactory derivation is simplified.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>